+++ server/src/start.py
@@ -1,22 +1,32 @@
 import zmq_connector
-import server
+from controller import Controller
+from db_connector import Db
+from server import Server
 
 
 class Start:
     
     def __init__(self):
-        self.my_server = None
-        self.mq = None
-    
-    def server(self):
-        # connector admits host ip in constructor
-        self.my_server = server.PullPubServer(zmq_connector.ZmqConnector())
-        # self.my_server.mq = self.mq
+        # start database
+        self.db = Db()
         
-        self.my_server.set_up()
-        self.my_server.loop()
+        # create ZMQ connector
+        self.mq = zmq_connector.ZmqConnector()
+        
+        # start server
+        self.server = Server(self.mq, self.db)
+        
+        # start controller
+        # self.controller = Controller(self.mq)
+
+    # def server(self):
+    #     # connector admits host ip in constructor
+    #     self.my_server = server.PullPubServer(zmq_connector.ZmqConnector())
+    #     # self.my_server.mq = self.mq
+    #
+    #     self.set_up()
+    #     self.my_server.loop()
 
 
 if __name__ == '__main__':
     start = Start()
-    start.server()